reuse token selector screen
