+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -3,6 +3,8 @@ import 'package:cryptoplease/app/screens/authenticated/flow.dart';
 import 'package:cryptoplease/app/screens/authenticated/home_tabs_screen.dart';
 import 'package:cryptoplease/app/screens/authenticated/receive_flow/routes.dart';
 import 'package:cryptoplease/app/screens/authenticated/wallet_screen.dart';
+import 'package:cryptoplease/core/presentation/token_selector_screen.dart';
+import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/features/activities/presentation/activities_screen.dart';
 import 'package:cryptoplease/features/add_funds/presentation/routes.dart';
 import 'package:cryptoplease/features/app_lock/presentation/routes.dart';
@@ -33,6 +35,7 @@ const authenticatedFlowRoutes = AutoRoute<void>(
         CustomRoute<void>(page: ActivitiesScreen),
         CustomRoute<void>(page: SwapTokenFlowScreen, children: swapTokenRoutes),
         CustomRoute<void>(page: MenuScreen),
+        CustomRoute<Token>(page: TokenSelectorScreen),
       ],
     ),
     ...backupPhraseRoutes,
@@ -44,7 +47,7 @@ const authenticatedFlowRoutes = AutoRoute<void>(
     ...receiveFlowRoutes,
     ...addFundsRoutes,
     appLockSetupFlowRoutes,
-    AutoRoute<void>(page: ProfileScreen),
+    AutoRoute<void>(page: ProfileScreen, fullscreenDialog: true),
     AutoRoute<void>(page: EditProfileScreen),
     AutoRoute<void>(page: TermsOfServiceScreen),
     AutoRoute<void>(page: PrivacyPolicyScreen),
